--- react-todo-app/src/Components/AddTodoForm.js
@@ -1,43 +1,0 @@
-import React, { useState } from "react";
-import InputWithLabel from "./InputWithLabel";
-
-const tableName = "Default";
-function AddTodoForm({ onAddTodo }) {
-  const [todoTitle, setTodoTitle] = useState("");
-
-  const handleTitleChange = (e) => {
-    const newTodoTitle = e.target.value;
-    setTodoTitle(newTodoTitle);
-  };
-
-  const handleAddTodo = (e) => {
-    e.preventDefault();
-    console.log(`input field => ${todoTitle}`);
-    if (todoTitle.trim().length === 0) {
-      alert("Input field is empty");
-    } else {
-      onAddTodo({
-        id: Date.now(),
-
-        fields: {
-          Title: todoTitle,
-        },
-      });
-      setTodoTitle("");
-    }
-  };
-
-  return (
-    <form onSubmit={handleAddTodo}>
-      <InputWithLabel
-        todoTitle={todoTitle}
-        handleTitleChange={handleTitleChange}
-      >
-        Title:
-      </InputWithLabel>
-      <button type="submit">Add</button>
-    </form>
-  );
-}
-
-export default AddTodoForm;
