+++ react-todo-app/src/Components/AddTodoForm/AddTodoForm.module.css
@@ -1,0 +1,63 @@
+.formContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 20px;
+  backgroundcolor: "lightblue";
+}
+
+.form {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 10px;
+}
+
+.inputField {
+  flex: 1;
+  padding: 0.5rem 1rem;
+  border: none;
+  border-radius: 5px;
+  margin-right: 5px;
+  font-size: 16px;
+  height: 2rem; /* Set height to match button */
+  vertical-align: middle;
+}
+
+.addButton {
+  padding: 0.5rem 1rem;
+  border: none;
+  border-radius: 5px;
+  background-color: #4caf57;
+  color: white;
+  cursor: pointer;
+  line-height: 1em;
+  font-size: 16px;
+  vertical-align: middle;
+  width: 60px;
+  height: 1.5rem;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  line-height: 0.6em;
+  outline: 2px solid white;
+}
+
+.addButton:hover {
+  background-color: #8b8e3e;
+}
+
+.inputWithLabel {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  margin-bottom: 0.5rem;
+  font-size: 16px;
+  font-weight: bold;
+  margin-right: "8px";
+}
+
+.inputContainer {
+  display: flex;
+  align-items: center;
+}
