+++ react-todo-app/src/Components/TodoList/TodoList.js
@@ -1,9 +1,10 @@
 import React from "react";
-import TodoListItem from "./TodoListItem";
+import TodoListItem from "../TodoListItem/TodoListItem";
+import style from "./TodoList.module.css"
 
 function TodoList({ todoList, onRemoveTodo }) {
   return (
-    <ul>
+    <ul className={style.ul}>
       {todoList.map((todo) => {
         return (
           <TodoListItem onRemoveTodo={onRemoveTodo} key={todo.id} todo={todo} />
