+++ react-todo-app/src/Components/TodoList/TodoList.module.css
@@ -1,0 +1,9 @@
+/* .ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+}
+
+.ul li {
+  border-bottom: 1px solid #e2e2e2;
+} */
