--- react-todo-app/src/Components/TodoListItem.js
@@ -1,21 +1,0 @@
-import React from "react";
-
-function TodoListItem({ todo, onRemoveTodo }) {
-  return (
-    <li className="li" id={todo.id}>
-      {todo.fields.Title}
-
-      <button
-        className="rm"
-        onClick={() => {
-          onRemoveTodo(todo.id);
-        }}
-        type="button"
-      >
-        Remove
-      </button>
-    </li>
-  );
-}
-
-export default TodoListItem;
