+++ react-todo-app/src/Components/TodoListItem/TodoListItem.module.css
@@ -1,0 +1,25 @@
+.listItem {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.5rem 1rem;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  margin-bottom: 0.5rem;
+  width: 500px;
+  margin: 0 auto;
+  background-color: #dcefc9;
+}
+
+.button {
+  background: red;
+  border-radius: 50%;
+  color: #fff;
+  border: none;
+  justify-content: space-evenly;
+  outline: 2px solid rgb(183, 155, 196);
+}
+
+.button:hover {
+  background-color: #98c9e5;
+}
